refactor: integrate typescript and various dev dependencies
