+++ package.json
@@ -6,10 +6,10 @@
   "license": "MIT",
   "private": true,
   "engines": {
-    "yarn": ">=1.7.0",
-    "node": "=12.x"
+    "yarn": ">=1.7.0"
   },
   "scripts": {
+    "commit": "cz",
     "lint": "eslint .",
     "test": "jest",
     "start": "babel-node index.js"
@@ -19,14 +19,26 @@
     "@babel/core": "^7.1.0",
     "@babel/node": "^7.0.0",
     "@babel/preset-env": "^7.1.0",
+    "@types/jest": "^26.0.22",
+    "@types/node": "^14.14.41",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^9.0.0",
     "babel-jest": "^23.6.0",
+    "cz-conventional-changelog": "3.3.0",
     "eslint": "^5.6.0",
     "eslint-config-prettier": "^3.1.0",
     "eslint-plugin-prettier": "^2.6.2",
     "jest": "24.8.0",
+    "nodemon": "^2.0.7",
     "prettier": "^1.14.3",
-    "regenerator-runtime": "^0.12.1"
+    "regenerator-runtime": "^0.12.1",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.2.4"
+  },
+  "dependencies": {},
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-conventional-changelog"
+    }
   }
 }
