+++ src/index.ts
@@ -1,12 +1,12 @@
 import { Drug, Pharmacy } from "./pharmacy";
 
-import fs from "fs";
+import * as fs from "fs";
 
 const drugs = [
   new Drug("Doliprane", 20, 30),
   new Drug("Herbal Tea", 10, 5),
   new Drug("Fervex", 5, 40),
-  new Drug("Magic Pill", 15, 40)
+  new Drug("Magic Pill", 15, 40),
 ];
 const trial = new Pharmacy(drugs);
 
@@ -17,7 +17,7 @@ for (let elapsedDays = 0; elapsedDays < 30; elapsedDays++) {
 }
 
 /* eslint-disable no-console */
-fs.writeFile("output.txt", log, err => {
+fs.writeFile("output.txt", log, (err) => {
   if (err) {
     console.log("error");
   } else {
